Implement a sample TESLA assertion using libtesla for automata state storage:
  (assertion in VOP_WRITE)(vp, cred) -> previously(returned(
    mac_vnode_check_write(cred, vp), 0));

A simple automata implements the above, and a simple test bench checks a
number of cases to make sure that assertions are being properly enforced.
Mysteriously, it appears to work.  Ideally, these classes of assertions
would use per-thread storage, but I need to get that working in userspace
still.



svn:r119