Modify 'audit' example to show our Clang changes.
Note the bit of cheating in helper_which_asserts() (in syscalls.c): we currently don't actually parse __tesla_assert() blocks. Other than that, the Clang-generated instrumentation drives the automata as we expect it to.

In order for this example to compile now, you'll need a modified Clang (from github.com/trombonehero/clang, branch 'tesla'). Make a symlink to ${CLANG_BUILD_DIR}/lib and call it 'llvmlib'.

svn:r149