Stop the transfer file size appearing three times per line of ssh_data.dat
The transfer file size is needed for producing the axis labels, but it was
being output three times per line, rather than the once needed, e.g.

100MB 1.526145 0.134429100MB 1.556168 0.135364100MB 1.511124 0.163857

This did not make much difference since Gnuplot appears to ignore the
letters, and the difference in the number would be too small to be
noticable. With this patch, the output now looks like:

100MB 1.526145 0.134429 1.556168 0.135364 1.511124 0.163857

svn:r502