Add a simple minimalistic sample called "foo".
Implement everything that currently has to be done manually.  The
foo_assertion.c file also implements the optional feature for
private assertion error reporting trying to show the need for a
more information to be able to log the state transition to tell
what went wrong rather than to just barf.