Move the 'tesla' directory out of the way.
It would be great to be able to build the TESLA analysis and instrumentation code separately from LLVM and Clang. This will happen Very Soon Now, but I must first create a place for this code to live.

So: put the old TESLA include directory inside a directly called 'include'.