Do something sensible with parameters.
The purpose of keeping track of function call parameters is not to store their types: that is done by the FunctionDecl that we keep a pointer to. Rather, we will want to be able to test the equality of parameters against parameters used elsewhere in an assertion. For instance, when generating the pattern matching and automata lookup for:

assert(previously(foo(x) == y) || eventually(bar(x) == y))

it will probably be handy to know that we mean the same x and y, rather than just "argument to foo of type X" and "argument to bar of type X". That way, we can name the assertion as (__FILE__, __LINE__, __COUNTER__, context, x, y).