Simplify callee instrumentation.
We don't need to split blocks: it's quite simple (and supported) to insert instructions in the middle of existing blocks.

Using this technique should allow most instrumentation (though not callee instrumentation) to run in a BasicBlockPass, rather than a FunctionPass. That would make it easier for the LLVM optimisation manager to do its job.