Refactor TESLA representation code into library.
The data structures that represent TESLA assertions (sequences of events, names of functions, etc.) will be wanted in both analysis and instrumentation passes. This commit refactors the common code out into two libraries:

libTesla: actual in-memory representations
libTeslaClang: code to parse Clang AST data into TESLA representations

All of the representations in libTesla can describe themselves in YAML, which the instrumenter will be able to parse. That parsing is the Next Step.