Start to add some debug-build-only error checking.
When we call tesla_instance_put(), the instance ought to belong to the specified tesla_state ("class"). When running a debug build, add assertions to check that this is the case.

This should only happen in debug builds because it is potentially expensive (a linear walk of a class' instances), but I'll definitely want some "safety nets" like this when running auto-generated automata management code!