Allow parsing 'ANY' arguments.
Sometimes, we enter or leave a function and don't care what its arguments were, but we still might like to keep track of the number and names of its arguments. Allow ParseArgument() to be called in a variant that records this information but says "this argument is an Any; I don't care what its value is".