Convert assertions to real instrumentation calls.
This new code creates calls to realisable instrumentation functions: we pass all of the dynamic information that we need to look up automata.

For instance, the zero'th automaton will create an instrumentation function called __tesla_instrumentation_assertion_reached_0() that knows how to interpret the arguments expected by automaton 0.

This does break the 'strawman' build, as we are no longer calling the same function for all automata, but that code should follow very soon.