Create (and use) a __tesla_ignore event.
Sometimes, we want to express things like:

TSEQUENCE(
  x->bar = baz,
  foo(x) == y,
  entered(baz)
);

If we are to have a somewhat readable syntax (e.g. foo(x) == y rather than __tesla_return(foo(x), y)), we can't make a lot of guarantees about the types of the arguments to TSEQUENCE(). This is fine for the varargs, but before this commit, that means we can't put 'foo(x) == y' or 'x->bar = baz' as the *first* argument to TSEQUENCE: ISO C demands that we have a "real" first argument for __tesla_sequence(), and that argument must have a type.

By creating a static __tesla_event called '__tesla_ignore', we can make our TSEQUENCE() macro expand to __tesla_sequence(__tesla_ignore, __VA_ARGS__). This lets us use the kind of syntax described above.