Collapse some TESLA IR classes together.
The distinction between Expression and Event, as well as Sequence and Repetition, was silly and got in the way of arbitrary expressivity. It was also starting to dramatically increase the complexity of parsing, so let's get rid of the silliness and just have one Expression class and one Sequence class.

Unfortunately, this change was very wide-ranging; I had to touch almost everything.