Explicit representation of transitions.
Create struct tesla_transition and struct tesla_transitions. THese describe transitions that can be taken as the result of an event. The idea is that exactly one of these transitions should be taken.

This commit doesn't actually *use* tesla_transition, just defines the structure and a helper function that can print a human-readable version (for debugging).