Use the right ConstructKey()!
Don't try to construct a tesla_key out of whatever arguments we happen to have prepared for tesla_update_state... we are supposed to be passing the instrumentation function's arguments to ConstructKey along with a description of what they mean.

This requires modifications to the callee.ll test: we need to use the correct number of arguments for each function!