Re-enable inclusive-or assertions.
Thanks to kg365, we now implement inclusive-or properly, so let's turn the strawman's use of them back on!

This causes a failure when running the strawman demo: it seems that we don't properly conflate the (5->7) transition with (28->39) and whatever else cares about returned(security_check). We should fix this tomorrow.

Still, this makes for a most impressive-looking PDF of the FSA!