Work around Linux include directory problem.
On Linux, it seems that the 'clang' driver and the Clang tooling framework don't expand their cflags in the same way. In particular, the 'clang' driver passes -resource-dir ${LLVM_LIB_DIR}/clang/${CLANG_VERSION} to clang -cc1; it seems that we need to pass it explicitly to the TESLA analyser, too (but only on Linux!).

In the longer term, we should figure out what's going on here and fix it upstream. In the meantime, this workaround... works.