Further update to lit-based unit tests.
Use more regex and more test lines to ensure that, for instance, we have an unbroken chain of expected state transitions (x -> y -> z for some value of x,y,z). Previously, we basically tested that almost-hardcoded values were passed to tesla_update_state(), which didn't allow for TESLA to renumber the states, and we often didn't check that the relevant transitions were actually taken.

Where we did check that transitions were taken, it was usually of the form "something was updated", rather than "the automaton instance named above was updated from state X to state Y, where X and Y are defined in the tesla_update_state() output preamble".