Add tesla_transition.to_mask.
Sometimes, we need to specify that a transition should ignore certain keys. For instance, the TESLA event:

foo(x,y) == z

may be relevant to transitions that care about (x,y,z) as well as transitions that only care about a subset of these (for example, the "just stay in the same state and ignore that event" transition).

In order to support this, each transition should specify a mask for the name an instance should have after taking the transition.