Stop taking surrounding macro as constant's name.
This requires a very strange trick because the API functions that apparently do what we want produce inconsistent results (e.g. they give one result for TCPS_FIN_WAIT_1 and another one for TCPS_FIN_WAIT_2, despite these two constants being #defined in the same way, in the same file, within a few lines of each other).

I'd like to understand SourceLocation magic better in the medium term, but for now, this makes automata representations easier to use when debugging (and it fixes an issue in Trac).