More robust find-pointer-to-TESLA-var logic.
On current versions of LLVM, the address of non-pointer variable %foo will be stored in %foo.addr, so that's what we currently use to find the pointer to a TESLA-relevant variable. We really shouldn't rely on this convention, however: in older versions of LLVM, such as that installed in /usr/local on zenith, the address-of-a-variable is stored in a numbered register. There's not reason why this convention mightn't change in the future, either.

This commit provides new logic to explicitly look for variables that are locally allocated to store a pointer to a variable that we care about, irrespective of the former's name. This works on older LLVM installs, but it's also more robust against future variable name changes.