tesla_class_reset should not unlock a tesla_class.
tesla_class_reset() should only be called with a tesla_class that is already locked (for classes that require synchronisation); unlocking it is the responsibility of the code that took out the lock. With the current implementation, tesla_update_state() does a double release whenever it resets a class.

It's unfortunate that neither the Mac OS X nor the Linux pthreads implementations caught this problem! Only FreeBSD's implementation returned EPERM when releasing an unheld lock (double release).