Use llvm-config --libdir to find libraries.
Use llvm-config --includedir to find headers and only if we are using a
clang/llvm build tree directly add internal include directories to the
path.

With these changes it is possible to build tesla against an installed
llvm instance that isn't in the default include path.  This should also
reduce the risk of header leakage when the llvm instance is installed
and a build directory also exists.