Add test binaries that repeatedly create a few states and classes.
These are split into ones that call the low-level `Tesla_internal`
module, and others that invoke the higher-level `Tesla` equivalents.

They regularly call `Gc.compact` to spot any potential memory
corruption.  This is useful to run with the `runtime-variant d`
debug runtime, although this is not activated by default due to
the verbosity of the resulting output.