Make the compiler enumerate (begin,end) pairs.
Reinstate the tesla_lifetime structure, but this time, let it be a simple combination of a begin and end tesla_lifetime_event.

Instead of keeping init- and cleanup-centric hash tables that point to possibly-partially-initialised initstate objects, simply keep a vector of tesla_lifetime objects that have been generated by the TESLA compiler parts.

Note: this commit updates libtesla but not the compiler parts!