TranslationFn::AddInstr*(Automaton, FunctionEvent)
This method adds the instrumentation required by a FunctionEvent (currently C-only, but I think I've annotated the places where Objective-C support code needs to be inserted).

This commit splits AddInstrumentation() into high-level methods that understand the semantics of program events and a low-level method that is agnostic to them but understands the details of translation function layout.