Make 'cpp' format checks more specific.
Clang 3.3's version of 'cpp' requires an '-o' flag for the output file, but Apple's current version of Clang (which, bizarrely, is called Clang 5.0) doesn't. So, when figuring out the right flags to pass to 'cpp', check for the string 'clang version 3.3' rather than just 'clang', which is also contained in the string 'Apple LLVM version 5.0 (clang-500.2.79) (based on LLVM 3.3svn)'.

This lets TESLA pass various cpp-dependent tests on OS X Mavericks.