Add a mask of free variables to tesla_key.
This will be used to tell automata instances with a particular name mask that an event occurred, even if we don't know what the names should be.

For instance, we can say "all automata with name (a,b,x,*)", where 'x' is "a value that we don't currently know" and '*' is "a value may or may not be set".