Refactor for more complex arguments.
Previously, we assumed that every C argument would translate into one LLVM IR argument. Depending on the calling convention, this might not be the case, so instead of passing an argument pointer to ParseArg() functions, pass a std::function that can be used to generate the next argument pointer. When we encounter special calling convention requirements, we will be able to call this std::function multiple times.

This requires proper C++11 support, not just -Wno-c++11-extensions.